% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subbytype.R
\name{subbytype}
\alias{subbytype}
\title{Subsets Data frame Based On Variable Types.}
\usage{
subbytype(df)
}
\arguments{
\item{df}{- Input Data frame We Wish To Subset.}
}
\value{
Returns List of Data frames.
}
\description{
Returns a list of 6 data frames.
             List's first element contains subset of all Factor variables of the input data frame.
             Second element contains subset of all numeric and Integer variables of the input data frame.
             Third element contains subset of all logical variables of the input data frame.
             Fourth element contains subset of all complex variables of the input data frame.
             Fifth element contains subset of all character variables of the input data frame.
             Sixth element contains subset of all raw variables of the input data frame.
}
\examples{
numv<-c(1,2,3)
chrv<-c("a","b","c")
df<-data.frame(numv,chrv)
subbytype(df)
}
\author{
"Sandip Kumar Gupta", "sandip_nitj@yahoo.co.in"
}
